Implement final endpoint changes in a distinct file
